make integration fuzz tests into unit fuzz tests for visibility reasons
