+++ runtime/near-vm-runner/src/tests/fuzzers.rs
@@ -1,8 +1,16 @@
 use arbitrary::Arbitrary;
+use bolero::check;
 use core::fmt;
 use near_primitives::contract::ContractCode;
-use near_vm_logic::VMContext;
-use near_vm_runner::internal::wasmparser::{Export, ExternalKind, Parser, Payload, TypeDef};
+use near_primitives::runtime::fees::RuntimeFeesConfig;
+use near_primitives::version::PROTOCOL_VERSION;
+use near_vm_errors::{FunctionCallError, VMError};
+use near_vm_logic::mocks::mock_external::MockedExternal;
+use near_vm_logic::{VMConfig, VMContext};
+
+use crate::internal::VMKind;
+use crate::internal::wasmparser::{Export, ExternalKind, Parser, Payload, TypeDef};
+use crate::VMResult;
 
 /// Finds a no-parameter exported function, something like `(func (export "entry-point"))`.
 pub fn find_entry_point(contract: &ContractCode) -> Option<String> {
@@ -89,4 +97,69 @@ impl fmt::Debug for ArbitraryModule {
         }
         Ok(())
     }
+}
+
+fn run_fuzz(code: &ContractCode, vm_kind: VMKind) -> VMResult {
+    let mut fake_external = MockedExternal::new();
+    let mut context = create_context(vec![]);
+    context.prepaid_gas = 10u64.pow(14);
+    let config = VMConfig::test();
+    let fees = RuntimeFeesConfig::test();
+
+    let promise_results = vec![];
+
+    let method_name = find_entry_point(code).unwrap_or_else(|| "main".to_string());
+    let res = vm_kind.runtime(config).unwrap().run(
+        code,
+        &method_name,
+        &mut fake_external,
+        context,
+        &fees,
+        &promise_results,
+        PROTOCOL_VERSION,
+        None,
+    );
+
+    // Remove the VMError message details as they can differ between runtimes
+    // TODO: maybe there's actually things we could check for equality here too?
+    match res {
+        VMResult::Ok(err) => VMResult::Ok(err),
+        VMResult::Aborted(mut outcome, _err) => {
+            outcome.logs = vec!["[censored]".to_owned()];
+            VMResult::Aborted(
+                outcome,
+                VMError::FunctionCallError(FunctionCallError::Nondeterministic(
+                    "[censored]".to_owned(),
+                )),
+            )
+        }
+    }
+}
+
+#[test]
+fn current_vm_does_not_crash() {
+    check!().for_each(|data: &[u8]| {
+        let module = ArbitraryModule::arbitrary(&mut arbitrary::Unstructured::new(data));
+        let module = match module {
+            Ok(m) => m,
+            Err(_) => return,
+        };
+        let code = ContractCode::new(module.0.module.to_bytes(), None);
+        let _result = run_fuzz(&code, VMKind::for_protocol_version(PROTOCOL_VERSION));
+    });
+}
+
+#[test]
+fn wasmer2_and_wasmtime_agree() {
+    check!().for_each(|data: &[u8]| {
+        let module = ArbitraryModule::arbitrary(&mut arbitrary::Unstructured::new(data));
+        let module = match module {
+            Ok(m) => m,
+            Err(_) => return,
+        };
+        let code = ContractCode::new(module.0.module.to_bytes(), None);
+        let wasmer2 = run_fuzz(&code, VMKind::Wasmer2);
+        let wasmtime = run_fuzz(&code, VMKind::Wasmtime);
+        assert_eq!(wasmer2, wasmtime);
+    });
 }
